added working SearchList + child components
