+++ src/components/original/Panel.js
