+++ src/components/original/PanelControls.js
